Remove `require('crypto')` from generateID
